Crate sp_crypto_ec_utils

This crate offers elliptic curves types which are compatible with the Arkworks library functionalities.

The implementation has been primarily designed to be used in slow hosted targets (e.g. wasm32) and offloads the most computationally expensive operations to the host by leveraging the arkworks-extensions library and Substrate’s host functions.

The exported types are organized and named in a way that mirrors the structure of the types in the original Arkworks library. This design choice aims to make it easier for users already familiar with the library to understand and utilize the exported types effectively.

Modules

bls12_381

BLS12-381 types and host functions.