Module sp_crypto_ec_utils::ed_on_bls12_381_bandersnatch

Elliptic Curves host functions to handle some of the Arkworks Ed-on-BLS12-381-Bandersnatch computationally expensive operations.

Modules

host_calls

Interfaces for working with Arkworks Ed-on-BLS12-381-Bandersnatch elliptic curve related types from within the runtime.

Structs

HostHooks

Curve hooks jumping into host_calls host functions.

Type Aliases

BandersnatchConfig

Group configuration.

EdwardsAffine

Twisted Edwards form point affine representation.

EdwardsConfig

Group configuration for Twisted Edwards form (equal to BandersnatchConfig).

EdwardsProjective

Twisted Edwards form point projective representation.

SWAffine

Short Weierstrass form point affine representation.

SWConfig

Group configuration for Short Weierstrass form (equal to BandersnatchConfig).

SWProjective

Short Weierstrass form point projective representation.